Revamp list of columns shown and allow switching last column.
Showing the project root and the command in the same table is tricky because
both can be very long values. The table framework does not work well when more
than one cell in a row holds long values. So the / keybinding was added to
toggle the value shown in the last column.